requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
python-multipart==0.0.9
pypdf==4.0.1
sentence-transformers==2.3.1
groq==0.4.0
httpx==0.27.2
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
